- Added INT64 type tag - Renamed other tags and IR instructions for more consistency - Still need to fix set_value and load_u64, store_u64

+++ source/analysis/typeset.d
@@ -183,7 +183,7 @@ struct TypeSet
                 assert (false, "unknown const type");
             break;
 
-            case Type.FLOAT:
+            case Type.FLOAT64:
             flags = TYPE_INT | TYPE_FLOAT;
             rangeMin = word.floatVal;
             rangeMax = word.floatVal;
